fixed comments from unit2 - 06.02.2020  - get parameter from annotaion  - rename all tasks packages
