+++ unit2/src/main/java/ru/task1/Main.java
@@ -1,4 +1,4 @@
-package task7;
+package ru.task1;
 
 public class Main {
