+++ unit2/src/main/java/ru/task1/Pen.java
@@ -1,4 +1,4 @@
-package task1;
+package ru.task1;
 
 import java.util.Objects;
