+++ unit2/src/main/java/ru/task2/Stationery.java
@@ -1,6 +1,4 @@
-package task3;
-
-import java.util.Objects;
+package ru.task2;
 
 public class Stationery {
     private String name;
@@ -25,21 +23,6 @@ public class Stationery {
 
     public void setPrice(Integer price) {
         this.price = price;
-    }
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Stationery that = (Stationery) o;
-        return Objects.equals(name, that.name) &&
-                Objects.equals(price, that.price);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, price);
     }
 
     @Override
