+++ unit2/src/main/java/ru/task3/Main.java
@@ -1,4 +1,4 @@
-package task3;
+package ru.task3;
 
 import java.util.Arrays;
 import java.util.Comparator;
