+++ unit2/src/main/java/ru/task3/Pen.java
@@ -1,4 +1,4 @@
-package task3;
+package ru.task3;
 
 public class Pen extends Stationery {
