+++ unit2/src/main/java/ru/task3/Pencil.java
@@ -1,4 +1,4 @@
-package task3;
+package ru.task3;
 
 public class Pencil extends Stationery {
