+++ unit2/src/main/java/ru/task5/Main.java
@@ -1,4 +1,4 @@
-package task5;
+package ru.task5;
 
 import java.util.Arrays;
 import java.util.List;
