+++ unit2/src/main/java/ru/task5/Subject.java
@@ -1,4 +1,4 @@
-package task5;
+package ru.task5;
 
 public class Subject {
