+++ unit2/src/main/java/ru/task6/Main.java
@@ -1,0 +1,16 @@
+package ru.task6;
+
+public class Main {
+
+    public static void main(String[] args) throws NoSuchMethodException {
+        NuclearSubmarin nuclearSubmarin = new NuclearSubmarin();
+
+        nuclearSubmarin.fullSpeedAhead();
+        nuclearSubmarin.stopSubmarine();
+
+        System.out.println(NuclearSubmarin.class
+                .getDeclaredMethod("fullSpeedAhead")
+                .getAnnotation(Info.class)
+                .author());
+    }
+}
