+++ unit2/src/main/java/ru/task6/NuclearSubmarin.java
@@ -1,4 +1,4 @@
-package task6;
+package ru.task6;
 
 public class NuclearSubmarin {
