--- unit2/src/main/java/task6/Main.java
@@ -1,11 +1,0 @@
-package task6;
-
-public class Main {
-
-    public static void main(String[] args) {
-        NuclearSubmarin nuclearSubmarin = new NuclearSubmarin();
-
-        nuclearSubmarin.fullSpeedAhead();
-        nuclearSubmarin.stopSubmarine();
-    }
-}
